+++ CHANGELOG.md
@@ -36,6 +36,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
   `marginal_roi()`, `cpik()`, and `summary_metrics()`. This allows the use of
   forecasted data for analysis.
 * Add `optimize()` method to the `OptimizationGrid` dataclass.
+* Add a warning when the target constraint of flexible budget optimization is
+  not met.
 
 ## [1.0.5] - 2025-03-06